fix(MessagePayload): fix bug where messages get squashed
This commit rewrites the utility `payload2data` function, which
is internally used to serialise MessagePayload data, to not use an
object in the global namespace. In addition to this, it refactors the
underlying logic behind serialisation so as to not slow down logic with
unnecessary if statements.

In addition to this, the commit also introduces functionality to add
descriptions to MessageAttachments. While implementing this, I noticed
that form-data bodies were significantly wacky in their implementation,
so I introduced a new option to either append JSON bodies to the
`payload_json` field in the multipart form data, or to simply squash it
into the form data itself. Appending to payload_json is significantly
more performant than other options as it simply JSON.stringifies instead
of running a costly for loop.